Alter upgrade function to only get upgrades on a Monday at work

+++ fish/functions/upgrade.fish
@@ -11,33 +11,47 @@ function upgrade
 
     switch (uname -s)
 
-       case Darwin
-          echo "Detected MacOS. Running MacOS specific managers..."
+        case Darwin
+            echo "Detected MacOS. Running MacOS specific managers..."
 
-       case Linux
-          echo "Detected Linux. Running Linux specific managers..."
-          sudo apt-get update -q
-          sudo apt-get upgrade -qy
-          sudo apt-get autoremove -qy
-          snap refresh
+        case Linux
+            echo "Detected Linux. Running Linux specific managers..."
+            sudo apt-get update -q
+            sudo apt-get upgrade -qy
+            sudo apt-get autoremove -qy
+            snap refresh
 
-       case '*'
-         echo 'Unknown OS'
-         exit 1
+        case '*'
+            echo 'Unknown OS'
+            exit 1
     end
 
-    echo "Updating brew taps..."
-    brew upgrade
-    brew cleanup
-    brew autoremove
+    if _is_work
+        if test (date | cut -f 1 -d ' ') = Mon
+           # Just do one big upgrade at the start of the week
+            echo "Updating brew taps..."
+            brew upgrade
+            brew cleanup
+            brew autoremove
 
-    echo "Updating fisher packages..."
-    fisher update
+            echo "Updating fisher packages..."
+            fisher update
 
-    echo "Updating vim..."
-    nvim +PlugUpgrade +PlugUpdate +PlugClean +PlugDiff
+            echo "Updating vim..."
+            nvim +PlugUpgrade +PlugUpdate +PlugClean +PlugDiff
+        end
 
-   if _is_work
-       work-upgrade
-   end
+        work-upgrade
+    else
+        echo "Updating brew taps..."
+        brew upgrade
+        brew cleanup
+        brew autoremove
+
+        echo "Updating fisher packages..."
+        fisher update
+
+        echo "Updating vim..."
+        nvim +PlugUpgrade +PlugUpdate +PlugClean +PlugDiff
+    end
 end
